+++ src/main/java/com/evacipated/cardcrawl/modthespire/MTSClassPool.java
@@ -1,19 +1,50 @@
 package com.evacipated.cardcrawl.modthespire;
 
+import javassist.ClassPath;
 import javassist.ClassPool;
 import javassist.CtClass;
+import javassist.LoaderClassPath;
 
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 public class MTSClassPool extends ClassPool
 {
     private ClassLoader classLoader;
+    private List<ClassPath> classPaths = new ArrayList<>();
 
-    public MTSClassPool(ClassLoader classLoader)
+    public MTSClassPool(MTSClassLoader classLoader)
     {
-        this.classLoader = classLoader;
         appendSystemPath();
+        resetClassLoader(classLoader);
+    }
+
+    public void resetClassLoader(MTSClassLoader loader)
+    {
+        classLoader = loader;
+        for (ClassPath classPath : classPaths) {
+            removeClassPath(classPath);
+        }
+        classPaths.clear();
+
+        insertClassPath(new LoaderClassPath(loader));
+        loader.addStreamToClassPool(this);
+    }
+
+    @Override
+    public ClassPath insertClassPath(ClassPath cp)
+    {
+        classPaths.add(cp);
+        return super.insertClassPath(cp);
+    }
+
+    @Override
+    public ClassPath appendClassPath(ClassPath cp)
+    {
+        classPaths.add(cp);
+        return super.appendClassPath(cp);
     }
 
     @Override